+++ libs/analytics/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [3.12.0](https://github.com/cowprotocol/cowswap/compare/analytics-v3.11.2...analytics-v3.12.0) (2026-09-23)
+
+### Minor Changes
+
+- feat: show twaps in explorer (#8129)
+
+- feat(solana): apply user slippage from settings (#8164)
+
+- feat: add new prop for TWAP for EOA analytics (#8201)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/common-utils@3.17.0
+  - @cowprotocol/common-const@3.13.0
+  - @cowprotocol/common-hooks@3.13.0
+  - @cowprotocol/currency@1.6.0
+  - @cowprotocol/events@4.13.0
+  - @cowprotocol/types@4.11.0
+
 ## [3.11.2](https://github.com/cowprotocol/cowswap/compare/analytics-v3.11.1...analytics-v3.11.2) (2026-09-15)
 
 ### Patch Changes