+++ libs/snackbars/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.3.5](https://github.com/cowprotocol/cowswap/compare/snackbars-v2.3.4...snackbars-v2.3.5) (2026-09-23)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/ui@3.14.0
+  - @cowprotocol/common-hooks@3.13.0
+
 ## [2.3.4](https://github.com/cowprotocol/cowswap/compare/snackbars-v2.3.3...snackbars-v2.3.4) (2026-09-15)
 
 ### Patch Changes