+++ libs/widget-lib/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [4.8.0](https://github.com/cowprotocol/cowswap/compare/widget-lib-v4.7.0...widget-lib-v4.8.0) (2026-09-23)
+
+### Minor Changes
+
+- feat: show twaps in explorer (#8129)
+
+- feat(solana): apply user slippage from settings (#8164)
+
+### Patch Changes
+
+- fix(widget): remove integrator typeform links (#8170)
+
+- Updated dependencies []:
+  - @cowprotocol/events@4.13.0
+  - @cowprotocol/iframe-transport@2.3.8
+
 ## [4.7.0](https://github.com/cowprotocol/cowswap/compare/widget-lib-v4.6.0...widget-lib-v4.7.0) (2026-09-09)
 
 ### Minor Changes